master: Fix integer type of chunk reference counter
This commit fixes integer type of chunk reference counter, which was
erroneously truncated from uint32_t to uint8_t.

+++ src/master/chunks.cc
@@ -222,7 +222,7 @@ public:
 	}
 
 	// number of files this chunk belongs to
-	uint8_t fileCount() const {
+	uint32_t fileCount() const {
 		return goalCounters_.fileCount();
 	}
